refactor: renamed markdown utils
